Modifico o TAD para traballar con ints

+++ TAD/vectordinamico.c
@@ -7,13 +7,13 @@ typedef int TELEMENTO;
 
 typedef struct {
     TELEMENTO *datos;
-    unsigned long int tam;
+    int tam;
 } STVECTOR;
 
 typedef STVECTOR *vectorD;
 
 /*Función CreaVector: asigna memoria y devuelve la asignación al vector*/
-void CreaVector(vectorD *v1, unsigned long int tam1) {
+void CreaVector(vectorD *v1, int tam1) {
     //LiberaVector(v1);
     *v1 = (vectorD) malloc(sizeof (STVECTOR));
     (*v1)->datos = (TELEMENTO*) malloc(tam1 * sizeof (TELEMENTO));
@@ -21,7 +21,7 @@ void CreaVector(vectorD *v1, unsigned long int tam1) {
 }
 
 /*Función AsignaVector: Llena una posición del vector con un valor*/
-void AsignaVector(vectorD *v1, unsigned long int posicion, TELEMENTO valor) {
+void AsignaVector(vectorD *v1, int posicion, TELEMENTO valor) {
     *( (*v1) ->datos + posicion) = valor;
 }
 
@@ -43,13 +43,13 @@ void LiberaVector(vectorD *v1) {
     }
 }
 
-TELEMENTO Componentei(vectorD v1,unsigned long int posicion){
+TELEMENTO Componentei(vectorD v1,int posicion){
     if (!EsNulo(v1))
         if (posicion < v1->tam)
             return *(v1->datos+posicion);
 }
 
-unsigned long int tamano(vectorD v1){
+int tamano(vectorD v1){
     return v1->tam;
 }
 
@@ -69,13 +69,13 @@ void liberar(vectorD *vectorV1) {
     }
 }
 
-TELEMENTO recuperar(vectorD v1, unsigned long int posicion){
+TELEMENTO recuperar(vectorD v1, int posicion){
     if (!EsNulo(v1))
         if (posicion < v1->tam)
             return *(v1->datos+posicion);
 }
 
-unsigned long int longitudVector(vectorD v1){
+int longitudVector(vectorD v1){
     if (!EsNulo(v1)){
         return v1->tam;
     }
@@ -84,8 +84,8 @@ unsigned long int longitudVector(vectorD v1){
 
 void imprimir(vectorD v1){
     TELEMENTO j;
-    unsigned long int longitud = longitudVector(v1);
-    unsigned long int i;
+    int longitud = longitudVector(v1);
+    int i;
     for (i = 0; i<longitud; i++){
         j = recuperar(v1,i);
         printf("%d ",j);
